+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Remove unused MultiReaderIterator and multiReader method ([#667](https://github.com/opensearch-project/opensearch-hadoop/pull/667))
 
 ### Fixed
+- Fixed global refresh when using dynamic index patterns ([#324](https://github.com/opensearch-project/opensearch-hadoop/issues/324))
 - Fixed build failures when downloading Apache project dependencies (Hadoop, Hive, Spark) ([#595](https://github.com/opensearch-project/opensearch-hadoop/pull/595))
 
 ### Security